# api/src/docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Float API
  description: |
    API de gestion des demandes de suppression de données personnelles (RGPD).

    **Authentification** : les routes protégées nécessitent un cookie
    HttpOnly `accessToken` obtenu via `POST /api/v1/auth/login`.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Serveur de développement local
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT stocké en cookie HttpOnly, obtenu via POST /auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Message d'erreur lisible
        code:
          type: string
          description: Code machine de l'erreur
      required: [error]
    RemovalRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        brokerId:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
        status:
          type: string
          enum: [DRAFT, PENDING, SENT, ACKNOWLEDGED, COMPLETED, REFUSED, NO_RESPONSE, COMPLAINT, SUPPRESSED]
        parentRequestId:
          type: string
          format: uuid
          nullable: true
          description: Renseigné si cette demande est une relance d'une demande initiale
        scheduledAt:
          type: string
          format: date-time
          nullable: true
          description: Date d'envoi planifié (PENDING → SENT)
        sentAt:
          type: string
          format: date-time
          nullable: true
        respondedAt:
          type: string
          format: date-time
          nullable: true
        nextActionAt:
          type: string
          format: date-time
          nullable: true
        emailBody:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /requests:
    get:
      tags: [Requests]
      summary: Lister les demandes de l'utilisateur connecté
      description: |
        Retourne les demandes de suppression de l'utilisateur authentifié,
        avec filtres optionnels (statut, broker) et pagination.
      security:
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          required: false
          description: Filtrer par statut
          schema:
            type: string
            enum: [DRAFT, PENDING, SENT, ACKNOWLEDGED, COMPLETED, REFUSED, NO_RESPONSE, COMPLAINT, SUPPRESSED]
        - name: broker_id
          in: query
          required: false
          description: Filtrer par broker (UUID)
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          description: Numéro de page (défaut 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Résultats par page (défaut 10, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Liste paginée des demandes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/RemovalRequest'
                        - type: object
                          properties:
                            brokerName:
                              type: string
                              nullable: true
                            brokerUrl:
                              type: string
                              nullable: true
                            brokerCategory:
                              type: string
                              nullable: true
                  total:
                    type: integer
                    description: Nombre total de résultats (avant pagination)
                  page:
                    type: integer
                  limit:
                    type: integer
                  totalPages:
                    type: integer
              example:
                data: []
                total: 0
                page: 1
                limit: 10
                totalPages: 0
        '400':
          description: Paramètre de filtre invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_status:
                  summary: Statut inexistant
                  value:
                    error: "Statut invalide. Valeurs acceptées : DRAFT, PENDING, SENT, ACKNOWLEDGED, COMPLETED, REFUSED, NO_RESPONSE, COMPLAINT, SUPPRESSED"
                    code: BAD_REQUEST
                invalid_broker_id:
                  summary: broker_id non UUID
                  value:
                    error: broker_id doit être un UUID valide.
                    code: BAD_REQUEST
        '401':
          description: Cookie accessToken absent ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Unauthorized
        '500':
          description: Erreur interne
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Impossible de récupérer la liste des demandes.
                code: INTERNAL_SERVER_ERROR

    post:
      tags: [Requests]
      summary: Créer une demande (DRAFT)
      description: |
        Crée une nouvelle demande de suppression en statut **DRAFT** pour
        l'utilisateur connecté.

        Le corps de l'email (`emailBody`) est généré immédiatement en interpolant
        les variables du template avec les données réelles de l'utilisateur
        (déchiffrées), du broker et de la demande. Il est persisté pour traçabilité.

        L'adresse postale est lue depuis `user_contacts` (type `address`,
        `isPrimary = true`). Si elle est absente, la valeur `[Adresse non renseignée]`
        est utilisée.

        Un événement `created` est tracé automatiquement.

        La demande créée peut ensuite être prévisualisée (`GET /requests/{id}/preview`)
        puis envoyée (`POST /requests/{id}/send`).
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [brokerId, templateId]
              properties:
                brokerId:
                  type: string
                  format: uuid
                  description: UUID du broker ciblé
                templateId:
                  type: string
                  format: uuid
                  description: UUID du template email à utiliser
                scheduledAt:
                  type: string
                  format: date-time
                  nullable: true
                  description: Date d'envoi planifié optionnelle
            example:
              brokerId: "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
              templateId: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '201':
          description: Demande créée en statut DRAFT
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RemovalRequest'
              example:
                data:
                  id: "uuid-request"
                  userId: "uuid-user"
                  brokerId: "6ba7b810-9dad-11d1-80b4-00c04fd430c8"
                  templateId: "550e8400-e29b-41d4-a716-446655440000"
                  status: DRAFT
                  emailBody: "Madame, Monsieur,\n\nJe vous adresse..."
                  scheduledAt: null
                  sentAt: null
                  createdAt: "2026-06-10T09:00:00.000Z"
                  updatedAt: "2026-06-10T09:00:00.000Z"
        '400':
          description: Champ obligatoire manquant ou UUID invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missing_fields:
                  summary: brokerId ou templateId absent
                  value:
                    error: Les champs brokerId et templateId sont obligatoires.
                    code: BAD_REQUEST
                invalid_broker_id:
                  summary: brokerId non UUID
                  value:
                    error: brokerId doit être un UUID valide.
                    code: BAD_REQUEST
                invalid_template_id:
                  summary: templateId non UUID
                  value:
                    error: templateId doit être un UUID valide.
                    code: BAD_REQUEST
        '401':
          description: Cookie accessToken absent ou expiré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Unauthorized
        '404':
          description: Broker, template ou utilisateur introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                broker_not_found:
                  summary: Broker inexistant
                  value:
                    error: Le broker spécifié est introuvable.
                    code: NOT_FOUND
                template_not_found:
                  summary: Template inexistant
                  value:
                    error: Le template spécifié est introuvable.
                    code: NOT_FOUND
                user_not_found:
                  summary: Utilisateur inexistant
                  value:
                    error: Utilisateur introuvable.
                    code: NOT_FOUND
        '500':
          description: Erreur interne
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Impossible de créer la demande.
                code: INTERNAL_SERVER_ERROR
...
  /requests/{id}/preview:
    get:
      tags: [Requests]
      summary: Prévisualiser l'email d'une demande
      description: |
        Génère un aperçu de l'email (sujet + corps) en interpolant les variables
        du template avec les données réelles de l'utilisateur, du broker et
        de la demande.

        L'adresse postale est lue depuis `user_contacts` (type `address`, `isPrimary = true`).
        Si elle est absente, la valeur `[Adresse non renseignée]` est utilisée.
      parameters:
        - name: id
          in: path
          required: true
          description: UUID de la demande
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Email prévisualisé avec variables interpolées
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      subject:
                        type: string
                        description: Sujet de l'email avec variables remplacées
                      body:
                        type: string
                        description: Corps de l'email avec variables remplacées
              example:
                data:
                  subject: "Demande de suppression de données — Jean Dupont"
                  body: "Madame, Monsieur,\n\nJe vous adresse..."
        '400':
          description: UUID invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Format d'identifiant de demande invalide. Un UUID est attendu."
                code: BAD_REQUEST
        '404':
          description: Demande introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: La demande de suppression spécifiée est introuvable.
                code: NOT_FOUND
        '500':
          description: Erreur interne
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Une erreur interne est survenue lors de la prévisualisation de la demande.
                code: INTERNAL_SERVER_ERROR
...